finish TODO for pad supporting bool

+++ python/oneflow/nn/functional/functional_multi_head_attention.py
@@ -350,12 +350,14 @@ def multi_head_attention(
         v = flow.cat([v, bias_v.repeat(1, bsz, 1)])
         if attn_mask is not None:
             # TODO(WangYi): remove cast after https://github.com/Oneflow-Inc/oneflow/pull/9234 merged
-            attn_mask = pad(attn_mask.int(), (0, 1)).to(attn_mask.dtype)
+            # attn_mask = pad(attn_mask.int(), (0, 1)).to(attn_mask.dtype)
+            attn_mask = pad(attn_mask, (0, 1))
         if key_padding_mask is not None:
             # TODO(WangYi): remove cast after https://github.com/Oneflow-Inc/oneflow/pull/9234 merged
-            key_padding_mask = pad(key_padding_mask.int(), (0, 1)).to(
-                key_padding_mask.dtype
-            )
+            # key_padding_mask = pad(key_padding_mask.int(), (0, 1)).to(
+            #     key_padding_mask.dtype
+            # )
+            key_padding_mask = pad(key_padding_mask, (0, 1))
     else:
         assert bias_k is None
         assert bias_v is None
@@ -398,12 +400,14 @@ def multi_head_attention(
         )
         if attn_mask is not None:
             # TODO(WangYi): remove cast after https://github.com/Oneflow-Inc/oneflow/pull/9234 merged
-            attn_mask = pad(attn_mask.int(), (0, 1)).to(attn_mask.dtype)
+            # attn_mask = pad(attn_mask.int(), (0, 1)).to(attn_mask.dtype)
+            attn_mask = pad(attn_mask, (0, 1))
         if key_padding_mask is not None:
             # TODO(WangYi): remove cast after https://github.com/Oneflow-Inc/oneflow/pull/9234 merged
-            key_padding_mask = pad(key_padding_mask.int(), (0, 1)).to(
-                key_padding_mask.dtype
-            )
+            # key_padding_mask = pad(key_padding_mask.int(), (0, 1)).to(
+            #     key_padding_mask.dtype
+            # )
+            key_padding_mask = pad(key_padding_mask, (0, 1))
 
     # update source sequence length after adjustments
     src_len = k.size(1)
